Fix dtype kwarg in load_vlm_or_llm for newer transformers
Qwen3ForCausalLM.__init__() rejects 'dtype' as a keyword argument in
transformers >= 4.55. Use 'torch_dtype' which is the standard HF parameter
for from_pretrained().

+++ modelopt/torch/speculative/utils.py
@@ -528,7 +528,7 @@ def load_vlm_or_llm(
     model = model_cls.from_pretrained(
         model_name_or_path,
         trust_remote_code=trust_remote_code,
-        dtype=dtype,
+        torch_dtype=dtype,
         device_map=device_map,
         **extra,
     )
